refs #285: address PR review scheduler feedback
