+++ FoqosTests/SyncApplyServiceTests.swift
@@ -54,20 +54,6 @@ final class SyncApplyServiceTests: XCTestCase {
       emergencyManager: emergencyManager, deviceId: deviceId)
   }
 
-  @MainActor
-  private final class ManualProfileDeleteCommitScheduler {
-    private(set) var scheduledOperations: [@MainActor () -> Void] = []
-
-    func schedule(_ operation: @escaping @MainActor () -> Void) {
-      scheduledOperations.append(operation)
-    }
-
-    func runNext() {
-      let operation = scheduledOperations.removeFirst()
-      operation()
-    }
-  }
-
   private func makeService(
     scheduleProfileDeleteCommit: @escaping (@escaping @MainActor () -> Void) -> Void
   ) -> SyncApplyService {
