+++ docs/plans/2026-07-07-pr-289-review-r1.md
@@ -1,0 +1,109 @@
+# Address PR #289 Review Comments (Round 1)
+
+> **For Claude:** REQUIRED SUB-SKILL: Use superpowers:subagent-driven-development to implement this plan task-by-task.
+
+**Goal:** Address accepted and partially accepted review comments from PR #289.
+
+**Architecture:** Keep the two-phase profile delete design unchanged. Tighten the scheduler to the intended main-actor/yield semantics, make deferred-commit tests deterministic through a shared manual scheduler, and document the async commit failure behavior without adding new product-facing callback APIs.
+
+**Tech Stack:** Swift, SwiftUI, SwiftData, XCTest, CloudKit sync engine.
+
+---
+
+### Task 1: Normalize deferred scheduler test seams (comments #1, #2, #3)
+
+**Review comments addressed:**
+
+- #1 `Foqos/Models/BlockedProfiles.swift:503`: replace the GCD + Task double hop in `scheduleProfileDeleteCommit` with a pure main-actor task and explicit `Task.yield()`.
+- #2 `FoqosTests/SyncEngineControllerTests.swift:65`: remove the timing-dependent `Task.yield()` x10 drain from controller tests by injecting a deterministic scheduler.
+- #3 Sourcery general feedback: deduplicate `ManualProfileDeleteCommitScheduler` across test files.
+
+**Implementation steps:**
+
+1. Add shared test helper:
+   - New file: `FoqosTests/Helpers/ManualProfileDeleteCommitScheduler.swift`
+   - Move the existing manual scheduler shape from:
+     - `FoqosTests/MutationFunnelTests.swift:64-76`
+     - `FoqosTests/SyncApplyServiceTests.swift:57-69`
+   - Keep the type name `ManualProfileDeleteCommitScheduler` so existing call sites do not need renaming.
+   - The helper should be `@MainActor final class ManualProfileDeleteCommitScheduler`, not `private`, with:
+     - `private(set) var scheduledOperations: [@MainActor () -> Void] = []`
+     - `func schedule(_ operation: @escaping @MainActor () -> Void)`
+     - `func runNext()`
+
+2. Remove the duplicated nested scheduler classes from:
+   - `FoqosTests/MutationFunnelTests.swift`
+   - `FoqosTests/SyncApplyServiceTests.swift`
+
+3. Make `SyncEngineControllerTests` use the deterministic scheduler:
+   - Add a `ManualProfileDeleteCommitScheduler` property near the existing test harness properties.
+   - Initialize it in `setUp()`.
+   - Construct `SyncApplyService` with `scheduleProfileDeleteCommit: scheduler.schedule`.
+   - Remove the `drainDeferredProfileDeleteCommit()` helper.
+   - In `testGivenFetchedDeletion_WhenHandled_ThenTombstoneClearedAndPendingDeleteRemoved`, replace `await drainDeferredProfileDeleteCommit()` with `scheduler.runNext()`. If the test no longer awaits anything, make it synchronous.
+
+4. Change production scheduler:
+   - In `Foqos/Models/BlockedProfiles.swift`, replace:
+     ```swift
+     DispatchQueue.main.async {
+       Task { @MainActor in
+         operation()
+       }
+     }
+     ```
+     with:
+     ```swift
+     Task { @MainActor in
+       await Task.yield()
+       operation()
+     }
+     ```
+
+**Verification:**
+
+- Run focused tests:
+  ```bash
+  xcodebuild test -project FamilyFoqos.xcodeproj -scheme FamilyFoqos -destination 'platform=iOS Simulator,id=B9E4A679-BDF3-4541-A59F-DA4BE21F80ED' -parallel-testing-enabled NO -only-testing:FoqosTests/MutationFunnelTests -only-testing:FoqosTests/SyncApplyServiceTests -only-testing:FoqosTests/SyncEngineControllerTests | xcpretty
+  ```
+
+### Task 2: Document async deferred-delete commit semantics (comment #4)
+
+**Review comment addressed:**
+
+- #4 Sourcery general feedback: profile deletes now have asynchronous commit semantics; document or centralize this behavior rather than adding new hooks.
+
+**Implementation approach:**
+
+Do not add new callback/hook APIs. Add concise comments where future maintainers need the contract:
+
+1. In `Foqos/Models/BlockedProfiles.swift` above `scheduleProfileDeleteCommit`, document that:
+   - The helper intentionally returns before the durable save.
+   - Callers must handle commit failures inside the scheduled operation.
+   - This gives SwiftUI one main-actor turn to invalidate pending-deleted profile views before the save creates the post-save zombie window.
+
+2. In `Foqos/CloudKit/SyncEngine/MutationFunnel.swift` near the profile delete scheduled commit, document that:
+   - Local-originated delete failures after the caller returns are converted into tombstone cleanup, rollback, and logging.
+   - `.deleteRecord` is enqueued only after the deferred save and target-absence check succeed.
+
+3. In `Foqos/CloudKit/SyncEngine/SyncApplyService.swift` near the remote profile delete scheduled commit, document that:
+   - Remote apply commit failures are not UI-action failures.
+   - They are persisted as `failedApplies` and retried by the sync controller.
+   - The controller side effects run only after the deferred commit succeeds.
+
+**Verification:**
+
+- Run focused tests:
+  ```bash
+  xcodebuild test -project FamilyFoqos.xcodeproj -scheme FamilyFoqos -destination 'platform=iOS Simulator,id=B9E4A679-BDF3-4541-A59F-DA4BE21F80ED' -parallel-testing-enabled NO -only-testing:FoqosTests/MutationFunnelTests -only-testing:FoqosTests/SyncApplyServiceTests -only-testing:FoqosTests/SyncEngineControllerTests | xcpretty
+  ```
+
+### Final verification
+
+After both tasks:
+
+```bash
+swift-format lint --recursive .
+scripts/check-sync-guards.sh
+scripts/check-c2-guards.sh
+xcodebuild test -project FamilyFoqos.xcodeproj -scheme FamilyFoqos -destination 'platform=iOS Simulator,id=B9E4A679-BDF3-4541-A59F-DA4BE21F80ED' -parallel-testing-enabled NO -only-testing:FoqosTests/SafeModelViewTests -only-testing:FoqosTests/MutationFunnelTests -only-testing:FoqosTests/SyncApplyServiceTests -only-testing:FoqosTests/SyncEngineControllerTests -only-testing:FoqosTests/SyncEngineControllerCutoverTests | xcpretty
+```
